+++ Sources/StreamChatUI/ChatChannelList/ChatChannelListVC.swift
@@ -402,6 +402,7 @@ open class ChatChannelListVC: _ViewController,
             return
         }
         reloadChannels()
+        handleStateChanges(controller.state)
     }
 
     @available(*, deprecated, message: "Please use `filter` when initializing a `ChatChannelListController`")
@@ -417,16 +418,22 @@ open class ChatChannelListVC: _ViewController,
     // MARK: - DataControllerStateDelegate
 
     open func controller(_ controller: DataController, didChangeState state: DataController.State) {
+        handleStateChanges(state)
+    }
+
+    /// Called whenever the channels data changes or the controller.state changes.
+    /// It controls the visibility of the channel list state views.
+    open func handleStateChanges(_ newState: DataController.State) {
         if isChatChannelListStatesEnabled {
             var shouldHideEmptyView = true
             var isLoading = true
 
-            switch state {
+            switch newState {
             case .initialized, .localDataFetched:
-                isLoading = channels.isEmpty
+                isLoading = controller.channels.isEmpty
             case .remoteDataFetched:
                 isLoading = false
-                shouldHideEmptyView = !channels.isEmpty
+                shouldHideEmptyView = !controller.channels.isEmpty
             case .localDataFetchFailed, .remoteDataFetchFailed:
                 shouldHideEmptyView = emptyView.isHidden
                 isLoading = false
@@ -436,9 +443,9 @@ open class ChatChannelListVC: _ViewController,
             emptyView.isHidden = shouldHideEmptyView
             chatChannelListLoadingView.isHidden = !isLoading
         } else {
-            switch state {
+            switch newState {
             case .initialized, .localDataFetched:
-                if channels.isEmpty {
+                if controller.channels.isEmpty {
                     loadingIndicator.startAnimating()
                 } else {
                     loadingIndicator.stopAnimating()