+++ Tests/StreamChatTests/APIClient/Endpoints/Requests/MessageReactionRequestPayload_Tests.swift
@@ -24,7 +24,7 @@ final class MessageReactionRequestPayload_Tests: XCTestCase {
                 "type": payload.reaction.type.rawValue,
                 "score": payload.reaction.score,
                 "mood": "good one"
-            ]
+            ] as [String: Any]
         ])
     }
 }